Move HTTP-specific sigv4 code into its own module
