+++ aws/rust-runtime/aws-sigv4/src/event_stream.rs
@@ -1,0 +1,8 @@
+/*
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0.
+ */
+
+//! Utilities to sign Event Stream messages.
+
+// TODO: Future home of Event Stream signing
